10 mars 2023

+++ test_maven3/src/main/java/centuri/test_maven/Counter.java
@@ -4,25 +4,25 @@ import java.util.ArrayList;
 
 public class Counter {
 	
-	int sum_coordinate;
+	Double sum_coordinate;
 	
-	public Counter (int sum_coordinate) {
+	public Counter (Double sum_coordinate) {
 		this.sum_coordinate = sum_coordinate;
 	}
 	
-	public int get_Counter () {
+	public Double get_Counter () {
 		return this.sum_coordinate;
 	}
 	
-	public void set_Counter (int sum_coordinate) {
+	public void set_Counter (Double sum_coordinate) {
 		this.sum_coordinate=sum_coordinate;
 	}
 	
-	public void add_Counter(int coordinate) {
+	public void add_Counter(Double coordinate) {
 		this.sum_coordinate = this.sum_coordinate + coordinate;
 	}
 	
-	public void less_Counter(int coordinate) {
+	public void less_Counter(Double coordinate) {
 		this.sum_coordinate = this.sum_coordinate - coordinate;
 	}
